+++ client/src/App.tsx
@@ -78,33 +78,16 @@ function App() {
         <Overview
           product={product}
           reviewsMeta={reviewsMeta}
-          reviews={reviews}
         />
-        <List currentProduct={product} />
+        <List currentProduct={product}/>
         <QuestionsAndAnswers
           currProductId={product.id}
           currProductName={product.name}
         />
-        <div id="ratingsReviewsContainerId">
-          <ReviewsRatings />
-        </div>
+        <ReviewsRatings id="ratingsReviewsContainerId"/>
       </div>
     );
   }
-  return (
-    <div>
-      <Overview
-        product={product}
-        reviewsMeta={reviewsMeta}
-      />
-      <List />
-      <QuestionsAndAnswers
-        currProductId={product.id}
-        currProductName={product.name}
-      />
-      <ReviewsRatings id="ratingsReviewsContainerId"/>
-    </div>
-  );
 }
 
 export default App;
